move components aroudn
