+++ package.json
@@ -100,6 +100,7 @@
     "nock": "^7.0.2",
     "nodemon": "^1.8.1",
     "npm-check-updates": "2.5.8",
+    "redux-devtools": "^3.1.1",
     "style-loader": "^0.13.0",
     "supertest": "^1.1.0",
     "url-loader": "^0.5.7",
