+++ public/js/components/app.jsx
@@ -2,20 +2,13 @@ import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import store from '../store'
-import Todos from './todos'
-import NavBar from './navbar'
-import Footer from './footer'
+import Base from './base'
 
-let reactElement = document.getElementById('react')
+let reactElement = document.getElementById('react');
+
 render(
   <Provider store={store}>
-    <NavBar />
-    <Todos />
-    <Footer />
+    <Base />
   </Provider>,
   reactElement
 )
-
-//document.body.style["margin-bottom"] = document.getElementById('footer').clientHeight;
-//tooltips();
-//formatHtml();
